% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_title_comment.R
\name{comment_out}
\alias{comment_out}
\title{Comment out codes based on their languages.}
\usage{
comment_out(x, engine, extra_syntax = character(0L))
}
\arguments{
\item{x}{A character vector}

\item{engine}{A string that defines the engine of the \code{x}.
See \code{names(comment_syntax)} for the predefined engines.}

\item{extra_syntax}{A named character vector which defines extra syntax to comment out.
Each values should contain \verb{\%s} which is the placeholder to insert \code{x}.
If definitions collides among \code{comment_syntax} and \code{extra_syntax},
then the latter has priority.}
}
\description{
Comment out codes based on their languages.
}
\examples{
# Default behaviors
comment_out("R language", "R")
comment_out("Python language", "python")

# A customized behavior
comment_out(
  "Python language",
  "python",
  extra_syntax = c("python" = '"""\%s"""')
)

}
\seealso{
\code{\link[=comment_syntax]{comment_syntax()}}
}
