% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_title_comment.R
\docType{data}
\name{comment_syntax}
\alias{comment_syntax}
\title{Comment syntax.}
\format{
An object of class \code{character} of length 38.
}
\usage{
comment_syntax
}
\description{
A named vector which defines comment syntax. This is internally used by
\code{hook_title_comment()}, \code{comment_title()}, and \code{comment_out()}.
}
\keyword{datasets}
