% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_title_comment.R
\name{comment_title}
\alias{comment_title}
\title{Comment title to chunk code}
\usage{
comment_title(
  options = list(engine = "R"),
  template = "{engine}",
  extra_syntax = character(0L),
  ...
)
}
\arguments{
\item{options}{A list of chunk options.}

\item{template}{A template of title processed by \code{glue::glue()}.
The processed value is automatically commented out by \code{comment_out()}.}

\item{extra_syntax}{A named character vector which defines extra syntax to comment out.
Each values should contain \verb{\%s} which is the placeholder to insert \code{x}.
If definitions collides among \code{comment_syntax} and \code{extra_syntax},
then the latter has priority.}

\item{...}{Arguments passed to \code{glue::glue()}}
}
\description{
Comment title to chunk code
}
\examples{
comment_title(
  options = list(engine = "R"),
  template = "{engine}"
)

comment_title(
  options = list(engine = "css", label = "mystyle"),
  template = "{label}.{engine}"
)

}
\seealso{
\code{\link[=comment_out]{comment_out()}}
}
