% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_number_lines.R
\name{hook_numberLines}
\alias{hook_numberLines}
\title{Number lines on code blocks}
\usage{
hook_numberLines(targets = "source", .set = TRUE)
}
\arguments{
\item{targets}{A character vector specifying what kind of code blocks
automatically number lines (default: \code{"source"}). Choices are
\code{"source"}, \code{"output"}, \code{"message"}, \code{"warning"}, and/or \code{"error"}.\verb{}NULL` disables the automation.}

\item{.set}{\code{TRUE} or \code{FALSE} to set the hook.}
}
\value{
invisible hook function
}
\description{
Number lines on code blocks created by chunks, i.e. source, output,
message, warning, and/or error.
}
\examples{
hook_numberLines("source")

}
