% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_title_comment.R
\name{hook_title_comment}
\alias{hook_title_comment}
\title{Prepend title comment to chunks}
\usage{
hook_title_comment(
  trigger = "title",
  default = "{engine}",
  args = list(),
  .set = TRUE
)
}
\arguments{
\item{trigger}{A name of chunk option triggers the hook.}

\item{default}{A string or function as a default value of the \code{trigger}ing chunk option.
If string, the value is given to \code{glue::glue}, and evaluates the expression
by setting chunk options as an environment, and then automatically commented out.
If function, it must have \code{options} parameter which receives chunk options,
and return a string.}

\item{args}{A list of arguments given to \code{glue::glue()} or callable \code{default}.}

\item{.set}{\code{TRUE} or \code{FALSE} to set a default value and a hook to the \code{trigger}.}
}
\description{
Prepend title comments to each chunks.
By default, title is the name of the corresponding chunk engines.
}
\examples{
# Set hook which comments engine names of each chunks at the beginning.
hook_title_comment()

# Customize format
# For R chunks, let title comment be something like "chunk-label.R".
# Otherwise, fall back to the default behavior.
hook_title_comment(
  default = function(options) {
    if (options$engine == "R") {
      return(sprintf("# \%s.\%s", options$label, options$engine))
    }
    return(comment_title(options, template = "{engine}"))
  }
)

}
\seealso{
\code{\link[=comment_title]{comment_title()}}, \code{\link[=comment_out]{comment_out()}}, \code{\link[=comment_syntax]{comment_syntax()}}
}
