% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_figure_unit.R
\name{hook_figure_unit}
\alias{hook_figure_unit}
\title{Change the unit of figure size}
\usage{
hook_figure_unit(unit = "mm")
}
\arguments{
\item{unit}{A string of an unit (default: "mm"). Available units follow.
\code{measurements::conv_unit_options$length}}
}
\value{
invisible \code{NULL}
}
\description{
By default, figure size of R Markdown is specified with inches.
This function changes the default unit.
}
\details{
As a side effect, \code{fig.retina} is set to \code{NULL} because of
\url{https://github.com/yihui/knitr/issues/1876}.
}
\examples{
hook_figure_unit('mm')

}
