% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_title_comment.R
\name{hook_title_comment}
\alias{hook_title_comment}
\title{Prepend title comment to chunks}
\usage{
hook_title_comment(
  trigger = "title",
  default = "#' {label}.{engine}",
  ...,
  .set = TRUE
)
}
\arguments{
\item{trigger}{A name of chunk option triggers the hook.}

\item{default}{A string or function as a default value of the \code{trigger}ing chunk option.
If string, the value is given to \code{glue::glue}, and evaluates the expression
by setting chunk options as an environment.
If function, it must have \code{options} parameter which receives chunk options,
and return a string.}

\item{...}{Additional arguments given to \code{glue::glue()} or callable \code{default}.}

\item{.set}{\code{TRUE} or \code{FALSE} to set a default value and a hook to the \code{trigger}.}
}
\description{
Title of chunks can be determined and prepended based on chunk options.
}
